Optimizer: register cache and constant propagation optimizeurs

+++ testfiles/13_function_return_types/06_double_return_expression.c
@@ -1,7 +1,7 @@
 double divide(double a, double b) {
     return a / b;
 }
-double main() {
+int main() {
     return divide(10.0, 2.5);
 }
